requirements.txt
selenium==4.6.0
pytest==7.2.2
faker==18.3.1
allure-pytest==2.13.1
pytest-xdist==3.2.1
requests==2.28.1
jsonschema==4.17.3